+++ packages/orq-rc/src/models/operations/getpromptversion.ts
@@ -75,7 +75,7 @@ export type GetPromptVersionResponseFormatType = ClosedEnum<
 
 export type GetPromptVersionResponseFormatJsonSchema = {
   name: string;
-  strict: boolean;
+  strict?: boolean | undefined;
   schema: { [k: string]: any };
 };
 
@@ -712,14 +712,14 @@ export const GetPromptVersionResponseFormatJsonSchema$inboundSchema: z.ZodType<
   unknown
 > = z.object({
   name: z.string(),
-  strict: z.boolean(),
+  strict: z.boolean().optional(),
   schema: z.record(z.any()),
 });
 
 /** @internal */
 export type GetPromptVersionResponseFormatJsonSchema$Outbound = {
   name: string;
-  strict: boolean;
+  strict?: boolean | undefined;
   schema: { [k: string]: any };
 };
 
@@ -730,7 +730,7 @@ export const GetPromptVersionResponseFormatJsonSchema$outboundSchema: z.ZodType<
   GetPromptVersionResponseFormatJsonSchema
 > = z.object({
   name: z.string(),
-  strict: z.boolean(),
+  strict: z.boolean().optional(),
   schema: z.record(z.any()),
 });