+++ packages/orq-rc/RELEASES.md
@@ -2149,3 +2149,13 @@ Based on:
 - [typescript v3.8.0-rc.21] packages/orq-rc
 ### Releases
 - [NPM v3.8.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.21 - packages/orq-rc
+
+## 2025-06-12 21:45:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.561.0 (2.628.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.8.0-rc.22] packages/orq-rc
+### Releases
+- [NPM v3.8.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/3.8.0-rc.22 - packages/orq-rc